add getUserByEmail

+++ model/user.go
@@ -1,7 +1,7 @@
 package model
 
 type User struct {
-	ID       int32  `json:"id" gorm:"primaryKey;autoIncrement"`
+	ID       int64  `json:"id" gorm:"primaryKey;autoIncrement"`
 	FullName string `json:"full_name" gorm:"column:full_name"`
 	Password string `json:"password"`
 	Email    string `json:"email" gorm:"unique"`
